+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Updated to game version 14.0.3 (Hotfix 4)
 * Temporarily disabled extension to `Material`
+<!-- this space intentionally left blank -->
+* Fixed commands `/whisper` and `/<color>` echoing
 
 # v0.2.0